+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- 2023-08-11 at 21h17 · [[PowerShell - Convert XLS to XLSX]]
 - 2023-08-11 at 20h49 · [[Batch - Run Docker Diagnostics Check]]
 - 2023-08-11 at 20h44 · [[_README]]
 - 2023-08-11 at 20h36 · [[_README]]
@@ -47,4 +48,3 @@
 - 2023-08-10 at 20h32 · [[Partitioning]]
 - 2023-08-10 at 20h32 · [[Sharding]]
 - 2023-08-10 at 20h31 · [[Materialized View]]
-- 2023-08-10 at 20h31 · [[Indexing]]